backend/requirements.txt
flask
flask-cors
flask-session
redis
kafka-python
azure-eventhub
azure-eventhub-checkpointstoreblob
mysql-connector-python
werkzeug
# OpenTelemetry 패키지들
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-mysql
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-kafka-python
opentelemetry-exporter-otlp-proto-http